ISPN-10373 Store/Loader Non blocking SPI
* Added in new SPI
* PersistenceManagerImpl utilizes only new SPI
* Add in adapter for old SPI to new without changes
* DummyInMemoryStore converted to new SPI
* Async store converted to new SPI
* Added in BaseNonBlockingStore to be used by stores

+++ persistence/remote/src/test/java/org/infinispan/persistence/remote/upgrade/HotRodUpgradeWithStoreTest.java
@@ -109,7 +109,7 @@ public class HotRodUpgradeWithStoreTest extends AbstractInfinispanTest {
       return pm.getStores(DummyInMemoryStore.class).iterator().next();
    }
 
-   private int storeSize(TestCluster cluster) {
+   private long storeSize(TestCluster cluster) {
       return getDummyStore(cluster).size();
    }
